rename ssrUtils -> ssr-utils, add notification
